python/uline/uline_api/requirements.txt
backports-abc==0.4
certifi==2016.9.26
psycopg2==2.6.2
redis==2.10.5
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.3
tornado==4.4.2
ujson==1.35
WTForms==2.1
pytest-tornado
bcrypt
futures
cryptography
celery==3.1.23